Build new Rails apps without having Ruby installed
In order to create a new Rails app in a docker container,
you have to have Ruby and Rails installed, run Rails new, and
then build the container.

This project allows the user to create a Rails app without having
Ruby installed. This is meant to be a portable way to create Rails
apps. The host machine does not need to have any software except
Bash and Docker, which you have to have to create a new Rails app
anyway, and are common on almost all systems.

Having a separate file in the porject violates the goal of this
project (to be a simple protable script). Therefore, the
Dockerfile should be stored as text in the bash script and
written at run time.