Replace --build option with --skip-build
Requiring the user to remember the build option is
a bad user experience. If the user passes in a Ruby
version that has never been built before and
forgets the build option, then the Rails app will
be built using whatever Ruby version was used
in the existing image.

The images also didn't have descriptive names, so
this commit adds that.